+++ libs/portals/my-pages/finance/src/lib/navigation.ts
@@ -1,5 +1,5 @@
 import { PortalNavigationItem } from '@island.is/portals/core'
-import { FJARSYSLAN_SLUG, HMS_SLUG, m } from '@island.is/portals/my-pages/core'
+import { m } from '@island.is/portals/my-pages/core'
 import { FinancePaths } from './paths'
 
 export const financeNavigation: PortalNavigationItem = {
