+++ libs/portals/my-pages/finance/src/screens/FinanceTransactionsVehicleMileage/index.ts
@@ -1,1 +1,1 @@
-export { default } from './FinanceTransactions'
+export { default } from './FinanceTransactionsVehicleMileage'
