+++ libs/portals/my-pages/finance/src/utils/filesHreyfingar.ts
@@ -1,6 +1,6 @@
 import { downloadFile } from '@island.is/portals/my-pages/core'
-import { CustomerRecordsDetails } from '../screens/FinanceTransactions/FinanceTransactionsData.types'
 import { hreyfingarHeaders } from './dataHeaders'
+import { CustomerRecordsDetails } from '../lib/types'
 const name = 'Hreyfingar'
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
